[refactor]: 필요 없는 코드 제거
